@@ -16,17 +16,17 @@
     "@babel/preset-env": "^7.24.4",
     "@babel/preset-react": "^7.24.1",
     "babel-loader": "^9.1.3",
-    "bootstrap": "^4.2.1",
+    "bootstrap": "^5.3.3",
     "fs-extra": "^7.0.1",
     "iframe-bridge": "^1.1.0",
     "js-yaml": "^3.12.1",
     "lockr": "^0.8.5",
     "lodash": "^4.17.21",
     "minimist": "^1.2.6",
-    "react": "^16.7.0",
-    "react-bootstrap": "^0.32.4",
-    "react-dom": "^16.7.0",
-    "react-visualizer": "^2.1.2",
+    "react": "^18.3.1",
+    "react-bootstrap": "^2.10.2",
+    "react-dom": "^18.3.1",
+    "react-visualizer": "^3.0.1",
     "superagent": "^4.1.0",
     "webpack": "^5.91.0"
   },
